spellbook/gitlab/agents/requirements.txt
fastapi
uvicorn
python-gitlab
openai
python-dotenv
pydantic
pyngrok
loguru
rich
argparse
